\name{assignment3Package-package}
\alias{assignment3Package-package}
\alias{assignment3Package}
\docType{package}
\title{
\packageTitle{assignment3Package}
}
\description{
\packageDescription{assignment3Package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{assignment3Package}
\packageIndices{assignment3Package}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{assignment3Package}

Maintainer: \packageMaintainer{assignment3Package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
