% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3_group11.R
\name{euclidean}
\alias{euclidean}
\title{Find greatest common divisor of two numbers.}
\usage{
euclidean(x1, x2)
}
\arguments{
\item{x1}{number 1}

\item{x2}{number 2}
}
\value{
The greatest common divisor of \code{x1} and \code{x2}
}
\description{
Find greatest common divisor of two numbers.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
