% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_calcs.R
\name{fish_calcs}
\alias{fish_calcs}
\title{A Fish Catch and Revenue Function}
\usage{
fish_calcs(table, priceTab, plot = TRUE)
}
\arguments{
\item{table}{Dataframe with Location as column names and Fish names as rows}

\item{priceTab}{Dataframe with columns for Fish name and Price}

\item{plot}{Do you want a plot of revenue across locations? Default is TRUE}
}
\description{
This function tells you which fish was most frequently caught per location, the revenue generated at each location, and plots the revenue across locations
}
\author{
Emma Siegfried
}
\keyword{fish}
\keyword{fish_calcs()}
