% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_catch.R
\docType{data}
\name{fish_catch}
\alias{fish_catch}
\title{Example Fish Catch Data}
\format{A data frame of 10 rows and 10 columns}
\usage{
fish_catch
}
\description{
Random sample of fish catch across locations
}
\author{
Emma Siegfried
}
\keyword{datasets}
