% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_prices.R
\docType{data}
\name{fish_prices}
\alias{fish_prices}
\title{Example Fish Price Data}
\format{A data frame of 10 rows and 2 columns}
\usage{
fish_prices
}
\description{
Random sample of price data per fish
}
\author{
Emma Siegfried
}
\keyword{datasets}
