% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_spc_growth.R
\name{fish_spc_growth}
\alias{fish_spc_growth}
\title{Fish: Specific Growth}
\usage{
fish_spc_growth(a = -0.497, b = 0.1656, c = 0.08588, d = -0.004266,
  temp_range)
}
\arguments{
\item{a}{Parameter a, Default is -0.4970}

\item{b}{Parameter b, Default is 0.1656}

\item{c}{Parameter c, Default is 0.08588}

\item{d}{Parameter d, Default is -0.004266}

\item{temp_range}{Water temperature (Celcius), Default is 22}
}
\value{
Specific Fish Growth
}
\description{
This function estimates specific fish growth as a function of temperature, using an equation from Bjoornsson et al., 2007. The function requires parameters a, b, c, d, as described in Bjoornsson et al., 2007, Table 2, and developed from geometric mean weights of cod
}
\references{
Björnsson, B., Steinarsson, A. & Árnason, T. Growth model for Atlantic cod (Gadus morhua): Effects of temperature and body weight on growth rate. Aquaculture 271, 216–226 (2007).
}
\author{
Emma Siegfried
}
