% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_parameters.R
\docType{data}
\name{growth_parameters}
\alias{growth_parameters}
\title{Example Fish Growth Data}
\format{A vector of 21 Celcius temperature values & Dataframe of one row and four columns
\itemize{
\item Range from 0 to 20 degrees Celcius
\item a Parameter, Set at -0.4970
\item b Parameter, Set at 0.1656
\item c Parameter, Set at 0.08588
\item d Parameter, Set at -0.004266
}}
\usage{
growth_parameters
}
\description{
Data from Bjoornsson et al., 2007, Table 2
}
\references{
Björnsson, B., Steinarsson, A. & Árnason, T. Growth model for Atlantic cod (Gadus morhua): Effects of temperature and body weight on growth rate. Aquaculture 271, 216–226 (2007).
}
\author{
Emma Siegfried
}
\keyword{datasets}
